feat: splitting the utils
