+++ menus/utils/KeyToolSignCallback.ts
@@ -1,0 +1,4 @@
+import { DidDocument, SignCallback } from '@kiltprotocol/sdk-js'
+
+
+export type KeyToolSignCallback = (didDocument: DidDocument) => SignCallback;
