+++ menus/utils/isCtypeOnChain.ts
@@ -1,0 +1,15 @@
+import { CType, ICType } from '@kiltprotocol/sdk-js'
+
+
+export async function isCtypeOnChain({
+  ctype,
+}: {
+  ctype: ICType;
+}): Promise<boolean> {
+  try {
+    await CType.verifyStored(ctype)
+    return true
+  } catch {
+    return false
+  }
+}
