+++ menus/utils/resolveOn.ts
@@ -1,0 +1,4 @@
+import { ChainHelpers } from '@kiltprotocol/sdk-js'
+
+
+export const resolveOn = ChainHelpers.Blockchain.IS_FINALIZED
